feat(MainSidebar): Add new files for type and demo data. #464
