+++ src/components/MainSidebar/icons/types.ts
@@ -1,3 +1,5 @@
+import { StaticImport } from "next/dist/shared/lib/get-img-props";
+
 export type IconsType = {
     color?: string
     width?: number
@@ -8,3 +10,16 @@ export type IconsType = {
     colorEllipseFour?: string
     colorEllipseThree?: string
 }
+
+export type Route = {
+    text: string;
+    path: string;
+    icon: React.ElementType;
+    navLinkTextColor?: string;
+    iconsColor?: string;
+};
+
+export type AppData = {
+    logo: { src: string | StaticImport; height: number; width: number };
+    url: string;
+};
